protoclient tweaks
